Reintroduce users need to specify XA_CONNECTION_PROPERTY variables
This follows some discussion and is not squashed with previous commits
in case we decide to undo some of this

This removes a test which checks the XA_CONNECTION_PROPERTY_XXX entries
previously determined by map_properties, which also did some validation
which no longer gets called. Since the names of these properties vary
by the database used it doesn't make sense to keep this validation.