Reintroduce the automagic setting of properties for mysql and postgresql datasources.
The driver name is either set in the:
- <PREFIX>_DRIVER envronment variable, or
- inferred from the DB_SERVICE_PREFIX_MAPPING entry (i.e. for 'internal'
  datasources only)

If we have both, the environment variable takes precedence

The automagic ONLY happens if the driver name equals 'mysql' or 'postgresql'. If either
of these names are used the driver MUST of course be the one indicated by the name.

For other names, the user does not get the automagic, similar to other datasources
such as oracle or db2.