Process errors and warnings before running the CLI script
as well as after. The reason is that when generating the CLI scripts
we sometimes need to do XPath checks which are running in a
sub-shell. When running in a sub-shell echo/log_error/log_warning
are not guaranteed to be shown in the console, so in these places
they should redirect to the cli error/warn file